parseTOML {RcppTOML}	R Documentation

Parse a TOML file

Description

The function parseTOML, also aliased to tomlparse and parseToml, parse TOML content inout. This is usually a file given by its name, or a character variable with a TOML expression.

Usage

parseTOML(input, verbose = FALSE, fromFile = TRUE, includize = FALSE,
  escape = TRUE)

tomlparse(...)

parseToml(...)

## S3 method for class 'toml'
print(x, ...)

## S3 method for class 'toml'
summary(object, ...)

Arguments

input
[character] TOML input, either as chracter value or path to TOML file

verbose
[logical] Optional verbosity flag, no current effect

fromFile
[logical] Optional with default value ‘TRUE’ indicating parsing from file

includize
[logical] Optional legacy option, no current effect

escape
[logical] Optional boolean whether strings should be escaped, defaults to ‘TRUE’

...
Used for aliased function and S3 methods

x
Used for S3 method

object
Used for S3 method

Details

This implementation uses the ‘tomlplusplus’ library by Mark Gillar (see https://github.com/marzer/tomlplusplus) which is TOML 1.0 compliant. In earlier work the package used the ‘cpptoml’ library by Chase Geigle (see https://github.com/skystrife/cpptoml for which development ceased once TOML 0.5.0 parity was reached.

Value

A list object with the parsed content as an S3 object of class ‘toml’

[Package RcppTOML version 0.2.3 Index]